add ntt v1

+++ kyber512/zeta_table.h
@@ -8,9 +8,9 @@
 #ifndef KYBER512_ZETA_TABLE_H_
 #define KYBER512_ZETA_TABLE_H_
 #include <stdint.h>
+#include "common.h"
 
-extern __device__ uint32_t zetaTab[];
-
+extern __device__ int16_t zetaTab[];
 
 
 #endif /* KYBER512_ZETA_TABLE_H_ */
